+++ validators/contactsValidator.js
@@ -1,0 +1,9 @@
+import Joi from "joi";
+
+const contactsSchema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().required(),
+  phone: Joi.number().required()
+}) 
+
+export default contactsSchema;
